requirements.txt
# Python==2.7.13
# Celery: Linux-Server: eg: CentOs

django>=1.11.5
pymongo>=3.4.0
mongoengine>=0.14.3

celery>=4.1.0
django-celery-results>=1.0.1
flower==0.9.2
redis>=2.10.6

djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0

requests>=2.12.4
openpyxl==2.4.8
xlsxwriter==0.9.8
python-scrapyd-api==2.0.1

chardet>=3.0.4

# django-allauth >= 0.32.0
# python-Levenshtein==0.12.0
